Avoid using global allocator when unpacking input
We SIGSYS on large inputs because we're calling `std::string` which uses
the default allocator. This is an attempt at solving the SIGSYS by
avoiding the allocation altogether.

We don't have to allocate a `std::string`: mruby copy the passed-in
buffer immediately. mruby is configured to use a non-global allocator
for which we've already allocated the maximum memory and is setup to not
make additional syscalls.